feat: esl-trigger rewritten
Get rid of base trigger that is not suit purposes
Fix bug of the keyboard handling
Rework API. Add ignore option.
